Instead of a change of not reproducing (qmat), only individuals which die will be replaced
BEFORE: Every individual had a chance of reproducing, however there was
a chance qmat that it wouldn't be able to do so. In that case, it was replaced
by another individual within its radius

IN THIS VERSION: There is a chance m (from mortality) that a individual will
die and be replaced by the offspring of one individual located within its
radius. Surviving individuals will remain the same.

PLEASE NOTE THAT NOW GENERATIONS ARE ONLY FULLY DISCRETE IF discretegen= T
AND m=1

Additionally:
* Minor corrections in README
* .Rhistory added to .gitignore
* Parameter name changed on R script (qmat -> m)